integrated with spi forecast fivethirtyeight

+++ src/main/kotlin/xyz/pierini/fantacombinator/model/input/Club.java
@@ -1,5 +1,7 @@
 package xyz.pierini.fantacombinator.model.input;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
@@ -9,5 +11,5 @@ import lombok.Setter;
 @AllArgsConstructor
 public class Club {
 	private String name;
-	private int weight;
+	private BigDecimal weight;
 }
