+++ src/main/kotlin/xyz/pierini/fantacombinator/model/output/AssociatedClub.java
@@ -1,5 +1,6 @@
 package xyz.pierini.fantacombinator.model.output;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import lombok.AllArgsConstructor;
@@ -13,7 +14,7 @@ public class AssociatedClub {
 
 	private String name;
 
-	private int associatedValue;
+	private BigDecimal associatedValue;
 	
 	private List<Integer> daysAgainstBigClub;
